Make set_affine_on_cr() a vfunc on RsvgRender
And use a new rsvg_drawing_ctx_set_affine_on_cr() throughout the code,
similar to the other vmethods accessed like
rsvg_drawing_ctx_get_pango_context().

This is a bit round-about, but it is because of this:  in the Rust
code we don't have direct access to RsvgRender; only to RsvgDrawingCtx
methods.  When we move all of that to Rust, this indirection will
disappear.