gitlabci: Limit cross-distos tests to only the GNOME/librsvg repo.
We currently have 3 test stages(not to be confused with the pipeline
stages).

Stage 1 is the relative cheap fedora builds and linting that
are run for every fork and every commit. It ensures that every commit
is buildable.

Stage 2 is the cross-distro and cross arch(amd64/i386/arm/etc) builds
that should will run on every commit in the master and stable branches
of the GNOME/librsvg repo.

Stage 3 are the expensive tests that should run only on schedules or
manually requested pipelines. Ex testing multiple gcc/rustc version
to ensure we don't accidently brake compatibility.

We do not want contributors to be confused if stage 1 passes but a later
stage 2 job fails due to probably a distro specific bug. Thus why stage 2+
is restricted to GNOME/librsvg unless a pipeline is explicitly manually
triggered from the Web GUI.

Previously that would be the case, but sometime people develop on fork/master
instead of a feature branch. I've done that too accidently many times. This
will make sure that such we won't have to worry about such occasions.