cond.rs: Use AsciiExt to maintain backwards compatibility.
In rustc 1.23 the "eq_ignore_ascii_case" method of the AsciiExt
was implemented directly for the types u8, chat, [u8] and str.
This causes the code to built fine with 1.23+ version of rustc
but failling for the v1.21 and v1.22 as the method does not exitst.

This patch imports the AsciiExt Trait and ignores the compiler
warning that would be produced when compiling with 1.23+ version
about an unused import.