Remove RsvgCairoClipRender; just use the current RsvgCairoRender
At this point, RsvgCairoClipRender was just a way to have a rendering
loop run (e.g. render a subtree of elements, but to build a clipping
path instead of normal rendering) without accumulating bounding boxes:
RsvgCairoClipRender had its own accumulator (the inherited ->bbox
field), but just ignored it after the clipping was done.

So, we just remove RsvgCairoClipRender and overwrite the bbox field in
the original RsvgCairoRender *with its original value* before we ran
the clipping loop.

We will be able to get rid of this when we can extract bounding boxes
outside of a full-blown rendering loop.