state: move dash array to rust
The Dasharray code in length.rs changed a bit: apart from dropping
the C functions, the struct now implements the Parse trait so that
we can use the property macros and does not handle "inherit" manually
anymore since the state code does it. The drawing code is in
draw.rs so that length.rs does not depend on cairo anymore.