state: move direction and writing mode to rust
Note that the logic used to set the direction on the pango context
is changed: in the C code, whichever of the two properties was set
last defined text_dir. Instead now we keep both in the state, then
when creating the pango layout we set the default dir in this way:
 - if bidi is set to override, from the direction prop (including
   its default)
 - otherwise if direction is specified, it wins
 - otherwise we default to the direction associated with the
   writing mode

I am not sure this is correct, but at least the logic is now in
one place, and it seems less aleatory than what was there before.
The test suite passes.