state: move the parsing of conditional processing attributes
Move (and slightly refactor) the parsing of conditional processing
attributes from rsvg_parse_presentation_attributes to its caller
rsvg_parse_style_attrs. This is not only cleaner (since the caller
is where we parse different categories of attributes), but it is
also more correct since rsvg_parse_presentation_attributes has
another caller in stop.rs and according to the spec that element
does not support conditional processing (one has to make the
whole gradient conditional, not just a gradient stop).