Remove RsvgRender base struct
All the code is already assuming that the only render is the cairo
render and casting back and forth. Once we are in rust we can bring
back a Render trait, but at the moment the "base" struct only gets
in the way.
While at it some code is also moved around: the draw_ctx creation
goes in base.c and the handle methods go in handle.c.