Replicate the old rsvg_state_clone() behavior with a little hack
That function would copy all the state's fields into the destination,
but preserve the destination's parent state.  We now do this by hand
in the single place where a state is cloned that way.

(In our Rust scheme of things, states are not meant to be
cloned... they are just clonable right now because of the "dominate"
nonsense.)