rsvg-load.c:node_set_atts(): Take in the real element's name instead of the one from the node creator
For unknown elements (e.g. "metadata", we use a default creator which
creates a NodeDefs.  However, this creator has a NULL element name,
since it is a placeholder for any unknown element.

So, now we pass the actual element name from the XML parser to
node_set_atts(), so that it can pass it in turn to the styles
machinery - it will need the name for proper CSS matching.