Don't have the affine in ComputedValues
THIS BREAKS A LOT OF STUFF.  The idea is that each Node carries its
own, un-aggregated transform.

I'm not yet sure where to carry the aggregated transform during
rendering.  Here I'm trying to put it implicitly in the draw_ctx.cr,
but this means we must save/restore the cr whenever we mess with its
transform.  Which makes sense in a Cairo-centric scheme of things, but
breaks our assumptions that the only way to change the cr's matrix is
through a hard cr.set_matrix().

Also, since we now use node.get_transform() where we previously used
values.affine (or even state.affine), things break because they are
now getting un-aggregated transforms.  I guess I'll have to audit this
at each step.