Node::draw() / draw_children() - Take a ComputedValues argument
It turns out that the <use> element acts as if it does a "deep clone
of the referenced element" but the styles from the referenced
element's ancestors are not used - the <use>'s styles are used
instead.

So, we'll need to have NodeUse do its own cascading for the purposes
of the referenced element: we will pass a ComputedValues to
Node::draw() and draw_children(); for NodeUse this will be a
custom-cascaded ComputedValues, and for normal rendering it will be
the values from the global cascade.