marker: Force display=inline for markers; not quite fixed yet
Markers are always displayed even if they or their ancestors are
display:none.  We force this and this fixes the case for ancestors
with display:none, but it doesn't work for markers with display:none
because after this point in the code, the toplevel loading calls
rsvg_parse_style_attrs() and overwrites our value.

Partially fixes painting-marker-07-f.svg

We may need some post-set-atts step...