property: rename method to 'compute' and take all values as param
Some properties need to know other properties' value in order to
compute their resulting value (e.g. baseline-shift needs to know
the parent font-size). Refactor all property macro machinery
to be parametric on the type of the ComputedValues type, rename
'inherit_from' to 'compute' and make it take the whole computed
values as param instead of just the computed value of the property
itself.
Note: the implementation of baseline-shift is not changed yet.