state_pop(): Removed.
state_push() is just used now to create the initial State in the
RsvgDrawingCtx, and get_current_state() is used in a couple of places
to fetch some values that we haven't redone yet.

This commit also makes state_push() only push the first time; all
subsequent calls don't add new states to the stack.