Regenerate tests that failed due to text rendering, after seeing what's going on
The old code would create/measure Pango layouts with the Cairo
context's affine set to an identity matrix.  It would later actually
render them with the final matrix.

The new code keeps the cr's matrix always updated, so the Pango
layouts get created for Cairo contexts that are already transformed.
This produces slight differences in rendering.  I think this version
is more correct, since the font machinery will have had a chance to
make decisions about hinting/etc. for the actual size at which the
text will be rendered, not an abstract unscaled size.