node::CascadedValues - return this from get_computed_values()
CascadedValues will turn into something similar to Cow.  It can either
deref to a node's own ComputedValues, or to a values that is cascaded
on demand.

The former will be for drawing elements in a normal context; they can
just use their precomputed cascade.  The latter will be for elements
being drawn in a <use> context, where the <use> overrides the cascade.