rsvg_drawing_ctx_draw_node_from_stack() - Pass an optional node to cascade from
According to the spec, elements referenced from feImage are suposed to
cascade like the <use> element.  The spec is not very clear on this,
but filters-composite-02-b.svg indicates that the referenced element
cascades from the original element that has the filter applied to it.
For example,

   <defs>
     <path id="foo" .../>

     <filter id="bar">
       <feImage ... xlink:href="#foo"/>
     </filter>
   <defs>

   <path id="baz" ... filter="url(#bar)"/>

Here, baz has a filter applied to it, and the feImage in the filter
references foo.  However, foo should inherit its CSS properties from
baz, not from bar.

This commit lets us give rsvg_drawing_ctx_draw_node_from_stack() a
second, optional node to cascade from.  If it is NULL, then the node
to draw will use its own default cascade.