(#273) - Pass the node being filtered down to the filter rendering functions
feImage wants to render the node it is referencing via href, but with
a cascade from the node to which the filter is being applied - similar
to the <use> element.  This requires that we are able to obtain the
node, to which the filter is being applied, from the filter rendering
function.

This commit introduces FilterContext::get_node_being_filtered().

pop_discrete_layer() also takes a node - this is so that it can create
the FilterContext with the node that is being filtered.

Also, Node::draw_children() now takes a node argument, so it can pass
it to pop_discrete_layer().

This makes filters-composite-02-b.svg render almost exactly like the
SVG 1.1 official reference image, except for the Arithmetic column;
the lower-right item has an extra row of pixels on top...

We regenerate filters-composite-02-b-ref.png for now.

https://gitlab.gnome.org/GNOME/librsvg/issues/273