Add tests for (#282), copied/regenerated from librsvg-2.42
(#282) - Don't reuse the input surface for painting feComposite's output

feComposite tried to take a shortcut if one of its inputs is an
existing surface, by painting its output to that same surface.
However, the overall filter may want to reuse the original surface
unchanged.  This was causing drop shadows, done like Inkscape's, to be
mis-painted.

https://gitlab.gnome.org/GNOME/librsvg/issues/282