Remove the with_layer argument throughout
Specifically, this was used in draw_children to decide whether to
push/pop a layer or not.

It was true everywhere, except in the two callers to
draw_in_viewport(), as that function already pushes a layer by itself.

So, to remove this exception and the with_layer argument, we now
explicitly push a layer wherever we need it.