viewport.rs: Inline everything into draw_in_viewport(); remove the tests (!)
The trait we used for mocking the RsvgDrawingCtx in the tests was
playing badly with the efforts to replace push/pop_discrete_layer()
with a single with_discrete_layer() function.

In the end I want to move to a model where RsvgDrawingCtx isn't
essentially a globally mutable mess, and intead functions pass around
a current Viewport that knows how to expose a cairo::Context and
things like the viewport size.

The tests here are removed, but they were essentially testing that we
call preserve_aspect_ratio() with the correct values.  I hope to have
more fine-grained tests in the future.