Use a RefCell for the vector of acquired nodes
get_acquired_node() is essentially a "give me a reference to a node",
with some runtime checking underneath.  Having AcquiredNode hold a
mutable reference to its parent DrawingCtx is highly inconvenient.
This calls for some interior mutability in DrawingCtx just for the
vector of acquired_nodes, so we use a RefCell.

Also, have AcquiredNode store a pointer to that RefCell, so that it
doesn't cause a borrow on the whole DrawingCtx.  This is unsafe, so
"impl Drop for AcquiredNode" has a little unsafe code.  Let's see if
we can refactor this in the future.