Don't keep a &DrawingCtx in FilterContext; pass it around to filter render functions
This changes a lot of code, because all filter rendering functions now
take a &mut DrawingCtx as an argument.  This not only makes them
consistent with the normal node rendering functions, but it lets us
avoid double mutable borrows if the draw_ctx needed to be kept in the
FilterContext.