Remove RsvgViewBox; return an additional gboolean from rsvg_node_svg_get_view_box()
The only purpose of the "active" field in

    typedef struct {
        cairo_rectangle_t rect;
        gboolean active;
    } RsvgViewBox;

was for rsvg_node_svg_get_view_box() to return whether the <svg>
element actually had a viewBox attribute.  If not, the .rect was
set to all-zeros.

We get rid of this struct, and return a cairo_rectangle_t directly
from that function plus an extra boolean.