gradient: take cairo tolerance into accont when fixing focus
Due to rounding errors, when we move the focus within the circle
we might end up slightly outside the circle and cairo will not
render the image correctly. We can get the tolerance from
the cairo context and make sure the focus point is inside
the circle within the given tolerance.
A reftest that fails before this patch is included.

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/54