gradient: take cairo tolerance into accont when fixing focus
Due to rounding errors, when we move the focus within the circle
we might end up slightly outside the circle and cairo will not
render the image correctly. We can get the tolerance from
the cairo context and take it into account when scaling.
A reftest that fails before this patch is included.
We also need to regenerate two of the existing reftests.

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/54