Node: add an element_name field
Normally each node type corresponds to a single element name, so in
theory we could derive the name from the NodeType.  However, we
translate unknown element names to NodeDefs, so a NodeDefs may refer
to not only <defs>.

Here we add an element_name field to the Node struct, so each specific
node may know to what SVG element it corresponds.