(#341): Don't infinite-loop with cyclic pattern references
In each loop iteration in resolve_pattern() we drop the AcquiredNode,
and so we can't use the implicit stack of acquired nodes to test for
cycles.

This commit introduces a NodeStack helper, which code can use to catch
reference cycles explicitly.

We also have new test program for files that cause infinite loops.

https://gitlab.gnome.org/GNOME/librsvg/issues/341