filters::render() - Take the computed values of the node being filtered, not the whole node
Some time soon I want to merge Artem Vorotnikov's work to replace the
ubiquitous Rc<Node> with a Rc<NodeTrait>, to have a "real" inheritance
scheme rather our awkward node.with_impl().

In the meantime, it turns out that the filters code doesn't need the
actual node that is being filtered; just a set of computed values to
pass around while rendering the filter.  So, let's do that.