(#347): Set the ink_rect of Pango layouts, not just their objectBoundingBox rect
The ink rectangle of Pango layouts was only being "grown" onto the
current bbox if the text was being rendered with a stroke.  We always
need a starting ink_rect, equal to the non-stroked bounds; we can grow
it later when needed.

This includes a new dimensions test for this issue.

https://gitlab.gnome.org/GNOME/librsvg/issues/347