css.rs: Fix usize cast
Previously we were casting usize, returned from the .len() method,
to a u64 and later convert that to libc::c_ulong in the C bindings.

But libc::ulong is u64 in amd64, but represented as u32 in i686,
which results into a type missmatch whe compiled in i686.

Givem that libc::c_ulong seems to behave like usize, it should be
safe cast usize to libc::c_ulong directly.