(#256): Fix matching of the systemLanguage attribute with the user's locale
The SVG spec says that the systemLanguage attribute is a
comma-separated list of BCP47 language tags, but librsvg was assuming
that it was a list of X/Open locale names.

Now, instead of fetching Unix-like names from g_get_language_names(),
we use the locale_config crate to get a list of BCP47 language tags.
Then, we use the language_tags crate to actually match those tags
against the values from the systemLanguage attribute.

https://gitlab.gnome.org/GNOME/librsvg/issues/256