(#256): Fix matching of the systemLanguage attribute with the user's locale
The SVG spec says that the systemLanguage attribute is a
comma-separated list of BCP47 language tags, but librsvg was assuming
that it was a list of X/Open locale names (e.g. what one would set in
the LC_MESSAGES or LANG environment variables).

Now, we convert the output of g_get_language_names() to BCP47 language
tags using the locale_config crate.

Then, we use the language_tags crate to actually match those tags
against the values from the systemLanguage attribute.

https://gitlab.gnome.org/GNOME/librsvg/issues/256