(#320) - Parse XmlLang as an identifier, not a CSS string
CSS strings are quoted, like "foo".  Our type for the xml:lang
attribute was being parsed as a CSS string, so an attribute like

  xml:lang="es"

would be parsed incorrectly, as the inner string would have had to be

  xml:lang="'es'"

So, we parse this attribute as an identifier now.

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/320