(#363): Fix whitespace removal around <tspan> elements
When there is something like

  <text>hello <tspan>world</tspan></text>

librsvg would trim away the space after "hello", so the rendered text
looked like "helloworld".

Now we only trim the ends of NodeChars elements if they don't have an
adjacent sibling; in the example above, the tspan is the sibling to
the right, and so the rightmost space in the "hello " NodeChars
doesn't get trimmed.

This makes the rendering in 340047.svg match Firefox; regenerated the
test reference file for that, and made it always use a sans-serif
font.

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/363