(#368): Log when an element is not rendered due to a nonexistent filter
This behavior changed since librsvg 2.40.  In the old version, if an
element references a nonexistent filter, the element would be rendered
unfiltered.

In the Rustified version of the filters code, the element appears to
not be rendered at all (in reality librsvg creates an empty filter
surface).  Firefox takes the same approach.

Having a message in the log when this happens will make it easier to
debug those invalid SVGs.

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/368