(#371): rsvg-convert - Fix positioning of extracted elements when they are scaled
When using the -w / -h options to set the width/height of the image,
together with the -i option to set the id of the single element to
export from an SVG, the resulting object was positioned incorrectly.
We want the object's upper-left corner to be at (0, 0) in the
resulting image.

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/371