(#372): Fix mis-rendering in small arc segments
The code was checking for division by 0.0 a bit too aggressively.  For
small radiuses (say, 0.25 units), this gives intermediate results that
are perfectly (and exactly!) representable within f64, but the code
was checking for 0 too coarsely.

Regenerated the test reference files for Adwaita that got minor
rendering changes because of this.

https://gitlab.gnome.org/GNOME/librsvg/issues/372