(#373): Render gradients with userSpaceOnUse even for an empty bounding box
In the test, a horizontal or vertical line was not being styled with a
gradient, because those lines have an empty bounding box.  However,
the gradient in question had gradientUnits="userSpaceOnUse", *not*
"objectBoundingBox".  So, we don't need the object's bounding box at
all, and the lines can use the gradient in the current coordinate
system.

Fixes https://gitlab.gnome.org/GNOME/librsvg/issues/373