Port data: url parsing to Rust
This uses the data-url crate.  One change in behavior is that data-url
is strict about data: URLs having a MIME-type chunk in them.  This
will be an apparent regression from issue #152, which requested that
image data be sniffed if a data: URL doesn't have a MIME-type.

However, https://fetch.spec.whatwg.org/#data-urls indeed indicates
that data: URLs without a MIME-type are invalid.

This removes the test files for issue #152, since they are expected to
fail now.