rsvg_decode_data_uri(): Restore ability to pass a null out_mime_type
Merge request !153 was due to a bug when calling
rsvg_decode_data_uri() with out_mime_type=NULL.  This was allowed in
the pre-rustification version of that function.  This commit restores
that behavior, and adds a test for

  <xi:include href="data:text/xml,&lt;xml...">

since the tests were not hitting that code path, which contained the crash.