Defs: decouple from RsvgHandle
Defs used to keep a pointer to its parent RsvgHandle from
initialization.  Instead, now it explicitly takes a handle into
.lookup().

This is so that we can create a Defs independently of an RsvgHandle
and manipulate it more freely.  Or maybe we can group it with all the
things that get produced at loading time, independently of a handle.

Unfortunately this means that the single call to defs.lookup() from
DrawingCtx now requires a handle, which DrawingCtx didn't have
before.  For now, we'll make DrawingCtx be constructed with a handle.
This more or less makes sense, since a DrawingCtx is really tied to a
particular handle.