Consolidate (tree, defs, css_styles) into an Svg type, invisible to C
The tree/defs/css_styles always go together, so we encapsulate them
into an Svg type.

In addition, we put this in the Rust Handle, not in the C
RsvgHandlePrivate, so the C code does not ever see the Svg type.

This requires moving a bunch of C-visible functions like
rsvg_defs_foo() to rsvg_handle_rust_defs_foo().  These will go away in
the end, as more of the code is ported to Rust.