Only handle internal general entities from libxml2
Entities which reference external resources are... problematic,
generally discouraged, and they don't seem to be used very much for
SVG, as it has mechanisms for referencing external files where it
needs them.

Also, this code was never correct; for example, the publicId is not an
URI that can be resolved from a base URI; it's just a string.

Don't handle external entities at all; just internal general (parsed)
entities, which libxml2 will substitute itself where they are used.