Export rsvg_handle_read_stream_sync
This is a litter weird, the CI seemed to have failed to build
due to the function not being exported, but locally it built
fine regardless of the environment and the rustc version. It
probably was a caching issue.

In any case, 9cb6c084653c232c63cfc17d4b95335f29a0b1c1 portted
rsvg_handle_read_stream_sync to Rust but did not export it
leading to the following linker error in the CI:

warning: function is marked #[no_mangle], but not exported
   --> rsvg_internals/src/handle.rs:646:1
    |
646 | / pub unsafe extern "C" fn rsvg_handle_rust_read_stream_sync(
647 | |     handle: *mut RsvgHandle,
648 | |     stream: *mut gio_sys::GInputStream,
649 | |     cancellable: *mut gio_sys::GCancellable,
...   |
671 | |     }
672 | | }
    | |_^
    |
    = note: #[warn(private_no_mangle_fns)] on by default
    = help: try exporting the item with a `pub use` statement

    Finished dev [unoptimized + debuginfo] target(s) in 1m 04s
  CCLD     librsvg-2.la

*** Warning: Linking the shared library librsvg-2.la against the
*** static library /builds/GNOME/librsvg/_build/target/debug/librsvg_internals.a is not portable!
  CCLD     rsvg-convert
/usr/bin/ld: ./.libs/librsvg-2.so: undefined reference to `rsvg_handle_rust_read_stream_sync'
collect2: error: ld returned 1 exit status