handle: move the whole fontmap hack to tests
There is no reason to set the default fontmap from within handle.
Setting the default fontmap is part of setting up the text fixture
and can live in test-utils. Beside we do not need to hold on to
the fontmap pointer, since setting by the default pango keeps its
own reference.