node: split resolve_resources method
Split a separate method in the trait that is used to load resources,
this allow to clean up the set_atts signature in most of the cases.
Later we can think of moving the actual load of the resources to
its own tree trasversal instead of doing it during parsing.