Handle::construct_read_stream_sync() - Move code from the exported function into here
Let's see how it works to have the from_glib_() conversions in the
pub extern functions, and to have the real work in Handle methods.

Also, this commit changes some functions to take borrowed references
to gio types (e.g. foo: &InputStream) instead of taking
ownership (foo: InputStream).  The former follows the gtk-rs
convention, and is the right thing anyway.