Remove defs.rs and move to the relevant files
The Defs struct goes into svg.rs, where it can be private and it is
renamed to Resources, since Defs is confusing (it has nothing to
do with NodeDefs anymore, since now we are storing ids in Svg itself).
The Href parsing code goes into allowed_url.rs.
Errors go to error.rs.