Support selecting a different Rust toolchain
At the moment the values passed through the RUSTC and CARGO environment
variables when running configure are ignored and just `rustc`/`cargo`
as found on the PATH are used.

This is problematic during package builds where we might have to
select a specific Rust toolchain, i.e., make sure to always use the
system Rust toolchain vs a rustup toolchain.

Signed-off-by: Tobias Kortkamp <tobik@FreeBSD.org>