Remove Parse's Data associated type.  Yay!
This removes the mysterious "()" argument from practically all the
.parse() and .parse_str() calls.

I thought the Data associated type was a really clever idea to deal
with Length's need for a direction.  But now it turns out that it's
possible to encode that in the length types.  In any case, it made the
parse() calls that didn't need extra data quite ugly.