render_to_pixbuf_at_size(): Handle a zero size by returning an empty pixbuf
SharedImageSurface: disallow creating with a zero-sized Cairo surface

gdk_pixbuf_get_file_info() uses a GdkPixbufLoader, but in its
"size-prepared" callback it saves the computed size, and then calls
gdk_pixbuf_loader_set_size(loader, 0, 0).  Presumably it does to tell
loaders that it only wanted to know the size, but that they shouldn't
decode or render the image to a pixbuf buffer.

Librsvg used to panic when getting (0, 0) from the size_callback:

- Handle.get_dimensions() would return (0, 0)

- Handle.get_pixbuf_sub() would create an ImageSurface of size (0, 0),
  which is valid in Cairo's terms.

- But SharedImageSurface used to do

    data_ptr = NonNull::new(cairo_image_surface_get_data(...)).unwrap()

  which panics when cairo_image_surface_get_data() returs NULL.  This
  is not even guaranteed; I think cairo just passes the (0, 0) down to
  pixman, and pixman ends up doing malloc(0), whose result is
  implementation-defined.  Thus, we can't guarantee that NonNull will
  work.  The caller has to be responsible for not creating zero-sized
  shared surfaces.