librsvg_crate: geometry_for_element() now takes a viewport
SVG really wants to be rendered/measured as *scalable*, based on the
viewport size you want it in.

This is especially important for SVGs which have percent dimensions:

  <svg width="100%" height="100%" viewBox=...>

In those, asking for the dimensions of an element doesn't make sense
unless there's a viewport size against which the "100%" will be
resolved.

In the C API, we resolved percent sizes to "SVG has no size".